requirements.txt
gradio==4.44.0
python-dateutil==2.8.2
openai==1.3.0
python-dotenv==1.0.0
